+++ main.go
@@ -58,7 +58,7 @@ type Agent struct {
 	CurrentTaskStatus api.TaskStatus
 	AutoUpdate        bool
 	Privileged        bool
-	NetworkHost       bool
+	NetworkMode       DockerNetworkMode
 }
 
 type Task api.TaskPollOutput
@@ -254,6 +254,7 @@ func (a *Agent) runWorker(ctx context.Context, task Task, taskStateChan chan tas
 	providedEnvVars := StringifyEnvironmentVariables(*task.WorkerEnvironmentVariables)
 	extraEnvVars := []string{
 		"RERUN_WORKER_ENVIRONMENT=dev",
+		fmt.Sprintf("RERUN_WORKER_DOCKER_NETWORK_MODE=%v", a.NetworkMode),
 	}
 	if a.Privileged {
 		extraEnvVars = append(extraEnvVars, "RERUN_WORKER_PRIVILEGED=true")
@@ -272,12 +273,6 @@ func (a *Agent) runWorker(ctx context.Context, task Task, taskStateChan chan tas
 	config := &container.Config{
 		Image: *task.WorkerImageURI,
 		Env:   append(providedEnvVars, extraEnvVars...),
-	}
-
-	var networkMode container.NetworkMode
-	networkMode = network.NetworkDefault
-	if a.NetworkHost {
-		networkMode = network.NetworkHost
 	}
 
 	res, err := a.Docker.ContainerCreate(
@@ -301,7 +296,6 @@ func (a *Agent) runWorker(ctx context.Context, task Task, taskStateChan chan tas
 					Target: "/root/.docker",
 				},
 			},
-			NetworkMode: networkMode,
 		},
 		&network.NetworkingConfig{},
 		&v1.Platform{},