+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Changes and release notes for the ReSim agent
 
 ## v0.2.4 - 2024-11-21
 
-- Added host-networking mode to enable running test workloads without network isolation
+- Added docker-network-mode to enable running test workloads with either bridge (default) or host networking mode
 
 ## v0.2.3 - 2024-11-18